+++ source/crisscross/core_io_writer.h
@@ -96,19 +96,19 @@ namespace CrissCross
 				 * \param _count The number of bytes to write.
 				 * \return The actual number of bytes written.
 				 */
-				virtual int WriteBlock(const void *_buffer, size_t _count);
+				virtual size_t WriteBlock(const void *_buffer, size_t _count);
 
 				/*! \brief Write an unsigned 8-bit integer. */
-				virtual int WriteU8(uint8_t _data);
+				virtual size_t WriteU8(uint8_t _data);
 
 				/*! \brief Write an unsigned 16-bit integer. */
-				virtual int WriteU16(uint16_t _data);
+				virtual size_t WriteU16(uint16_t _data);
 
 				/*! \brief Write an unsigned 32-bit integer. */
-				virtual int WriteU32(uint32_t _data);
+				virtual size_t WriteU32(uint32_t _data);
 
 				/*! \brief Write an unsigned 64-bit integer. */
-				virtual int WriteU64(uint64_t _data);
+				virtual size_t WriteU64(uint64_t _data);
 
 				/*! \brief Writes a string to the buffer. */
 				/*!
